modify testing structure in /api/server/test
